+++ example/a.ts
@@ -1,5 +1,23 @@
-import { Elysia, t } from '../src'
-import { hasTransform } from '../src/compose'
-import { post, req } from '../test/utils'
+import { Elysia, error, t } from '../src'
+import { req } from '../test/utils'
 
-import { treaty } from '../../eden'
+const app = new Elysia().get(
+	'/',
+	() => {
+		return {
+			name: 'a',
+			a: 'b'
+		}
+	},
+	{
+		response: {
+			200: t.Object({
+				name: t.String()
+			})
+		}
+	}
+)
+
+app.handle(req('/'))
+	.then((x) => x.json())
+	.then(console.log)